+++ docs-demo/start/Start.vue
@@ -1,0 +1,37 @@
+<script lang="ts" setup>
+import { ref, useTemplateRef, onMounted, onBeforeUnmount } from 'vue';
+import StkTable from '../StkTable.vue';
+import { StkTableColumn } from '../../src/StkTable/index';
+
+const stkTableRef = useTemplateRef('stkTableRef'); // vue3.5+
+// const stkTableRef = ref<InstanceType<typeof StkTable>>(); //vue3.2
+
+type DataType = {
+    id: string;
+    name: string;
+    age: number;
+};
+const columns: StkTableColumn<DataType>[] = [
+    { title: '姓名', dataIndex: 'name', key: 'name' },
+    { title: '年龄', dataIndex: 'age', key: 'age' },
+];
+const dataSource: DataType[] = [
+    { id: 'k1', name: '张三', age: 18 },
+    { id: 'k2', name: '李四', age: 19 },
+    { id: 'k3', name: '王五', age: 20 },
+];
+let interval = 0;
+onMounted(() => {
+    // 高亮指定的id行
+    interval = window.setInterval(() => {
+        stkTableRef.value?.setHighlightDimRow(['k1']);
+    }, 2000);
+});
+
+onBeforeUnmount(() => {
+    window.clearInterval(interval);
+});
+</script>
+<template>
+    <StkTable ref="stkTableRef" row-key="id" :columns="columns" :data-source="dataSource"></StkTable>
+</template>